Improve performance in tournament controller
Change selectedTeam from an array to an object in order to improve performance in the tournamentPlayerData function in the tournament controller. Searching for a player's id is faster in an object than iterating through an entire array. This search is made many times, so a check of the selectedPlayers object was added that will avoid unnecessary searches