Updating scores for a Tournament does not need to be run twice
When updating scores for a Tournament, the request had to be run twice when new Players were added. New Players are sometimes added while updating scores due to the source of the data not always being the same. This commit enables the admin to only have to update the request one time. New Players are now added to the specific Tournament table and then all the Players scores are updated at the same time